+++ cmd/traffic/cmd/manager/internal/mutator/agent_injector.go
@@ -93,7 +93,14 @@ func (a *agentInjector) inject(ctx context.Context, req *admission.AdmissionRequ
 	case "false", "disabled":
 		dlog.Debugf(ctx, `The %s.%s pod is explicitly disabled using a %q annotation; skipping`, pod.Name, pod.Namespace, agentconfig.InjectAnnotation)
 		return nil, nil
-	case "", "enabled":
+	case "":
+		if env.AgentInjectPolicy != agentconfig.OnDemand {
+			dlog.Debugf(ctx, `The %s.%s pod has not enabled %s container injection through %q annotation; skipping`,
+				pod.Name, pod.Namespace, agentconfig.ContainerName, agentconfig.InjectAnnotation)
+			return nil, nil
+		}
+		fallthrough
+	case "enabled":
 		config, err = a.findConfigMapValue(ctx, pod, nil)
 		if err != nil {
 			if isDelete {
@@ -104,8 +111,8 @@ func (a *agentInjector) inject(ctx context.Context, req *admission.AdmissionRequ
 		switch {
 		case config == nil && isDelete:
 			return nil, nil
-		case config == nil && ia == "":
-			dlog.Debugf(ctx, `The %s.%s pod has not enabled %s container injection through %q configmap or %q annotation; skipping`,
+		case config == nil && ia != "enabled":
+			dlog.Debugf(ctx, `The %s.%s pod has not enabled %s container injection through %q configmap or through %q annotation; skipping`,
 				pod.Name, pod.Namespace, agentconfig.ContainerName, agentconfig.ConfigMap, agentconfig.InjectAnnotation)
 			return nil, nil
 		case config != nil && config.Manual:
@@ -208,8 +215,9 @@ func addInitContainer(ctx context.Context, pod *core.Pod, config *agentconfig.Si
 		return patches
 	}
 
+	pis := pod.Spec.InitContainers
 	ic := agentconfig.InitContainer(config.AgentImage)
-	if len(pod.Spec.InitContainers) == 0 {
+	if len(pis) == 0 {
 		return append(patches, patchOperation{
 			Op:    "replace",
 			Path:  "/spec/initContainers",
@@ -217,7 +225,8 @@ func addInitContainer(ctx context.Context, pod *core.Pod, config *agentconfig.Si
 		})
 	}
 
-	for i, oc := range pod.Spec.InitContainers {
+	for i := range pis {
+		oc := &pis[i]
 		if ic.Name == oc.Name {
 			if ic.Image == oc.Image &&
 				slices.Equal(ic.Args, oc.Args) &&
@@ -299,7 +308,7 @@ func compareVolumeMounts(a, b []core.VolumeMount) bool {
 	stripKubeAPI := func(vs []core.VolumeMount) []core.VolumeMount {
 		ss := make([]core.VolumeMount, 0, len(vs))
 		for _, v := range vs {
-			if !strings.HasPrefix(v.Name, "kube-api-access-") {
+			if !(strings.HasPrefix(v.Name, "kube-api-access-") || strings.HasPrefix(v.MountPath, "/var/run/secrets/kubernetes.io/")) {
 				ss = append(ss, v)
 			}
 		}
@@ -314,7 +323,7 @@ func containerEqual(a, b *core.Container) bool {
 	return cmp.Equal(a, b,
 		cmp.Comparer(compareProbes),
 		cmp.Comparer(compareVolumeMounts),
-		cmpopts.IgnoreFields(core.Container{}, "ImagePullPolicy", "TerminationMessagePath", "TerminationMessagePolicy"))
+		cmpopts.IgnoreFields(core.Container{}, "ImagePullPolicy", "Resources", "TerminationMessagePath", "TerminationMessagePolicy"))
 }
 
 // addAgentContainer creates a patch operation to add the traffic-agent container