+++ integration_test/itest/connected.go
@@ -24,6 +24,7 @@ func WithConnection(np NamespacePair, f func(ctx context.Context, ch NamespacePa
 
 func (ch *connected) setup(ctx context.Context) bool {
 	t := getT(ctx)
+	_, _, _ = Telepresence(ctx, "quit", "-ur") //nolint:dogsled // don't care about any of the returns
 	// Start once with default user to ensure that the auto-installer can run OK.
 	stdout := TelepresenceOk(WithUser(ctx, "default"), "connect")
 	require.Contains(t, stdout, "Connected to context default")